Move npts, rotbaum; remove __init__ from test.
