+++ .github/workflows/tests-xgboost.yaml
@@ -28,4 +28,4 @@ jobs:
         pip install -r requirements/requirements-rotbaum-extra-methods.txt
     - name: Test with pytest
       run: |
-        pytest -n2 --doctest-modules test/model/rotbaum
+        pytest -n2 --doctest-modules test/ext/rotbaum
